Add basic code for rrepublican calendar
Create a class `Date::Republican` with an initializer for custom year, month, and day, and also an initializer `from_days` which receives an amount of days since the French republican epoch. It also opens the Date class to add the `to_republican` method to it (which returns a `Date::Republican` object. It has the reciprocal instance methods (`to_days` and `to_gregorian`)